LPFG-901: Moved RequestEntityFactory
