+++ src/main/java/uk/gov/cslearning/record/service/catalogue/LearningCatalogueService.java
@@ -6,7 +6,7 @@ import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
-import uk.gov.cslearning.record.csrs.service.RequestEntityFactory;
+import uk.gov.cslearning.record.service.RequestEntityFactory;
 
 import java.util.List;
